WIP Use cookies to protect route
This adds a cookie to the users session to confirm that they've
succeeded/failed at webauthn_verification and block the success/failure
pages when not appropriate.

Currently this is working in tests, but when tophatted locally the
cookie seems to delete and then come right back once we navigate back to
the success/failure page.